+++ backend/src/subscriptions/subscriptions.module.ts
@@ -1,0 +1,13 @@
+import { Module } from '@nestjs/common';
+import { TypeOrmModule } from '@nestjs/typeorm';
+import { SubscriptionsController } from './subscriptions.controller';
+import { SubscriptionsService } from './subscriptions.service';
+import { UsersModule } from '../users/users.module';
+import { Subscription } from '../models/subscription.entity';
+
+@Module({
+  imports: [UsersModule, TypeOrmModule.forFeature([Subscription])],
+  controllers: [SubscriptionsController],
+  providers: [SubscriptionsService],
+})
+export class SubscriptionsModule {}