+++ frontend/pages/admin/groups.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const GroupsPage = () => {
+  return (
+    <div>
+      <h1>Group Management</h1>
+      {/* Group management UI will go here */}
+    </div>
+  );
+};
+
+export default GroupsPage;