+++ frontend/pages/admin/users.js
@@ -1,0 +1,12 @@
+import React from 'react';
+
+const UsersPage = () => {
+  return (
+    <div>
+      <h1>User Management</h1>
+      {/* User management UI will go here */}
+    </div>
+  );
+};
+
+export default UsersPage;